backend/requirements.txt
azure-functions
azure-storage-blob>=12.14.0
pandas>=1.5.0
numpy>=1.23.0